# docs/openapi_contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Content Automation ERP API",
    "description": "Production-grade API-First Control Plane for YouTube Automation & Durable Workflows",
    "version": "1.0.0"
  },
  "paths": {
...
    "/seo/harvest/suggest": {
      "post": {
        "tags": [
          "SEO & Keywords"
        ],
        "summary": "Harvest Youtube Suggestions",
        "operationId": "harvest_youtube_suggestions_seo_harvest_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SEOHarvestSuggestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SEOHarvestSuggestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/seo/harvest/competitors": {
      "post": {
        "tags": [
          "SEO & Keywords"
        ],
        "summary": "Harvest Youtube Competitors",
        "operationId": "harvest_youtube_competitors_seo_harvest_competitors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SEOHarvestCompetitorsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SEOHarvestCompetitorsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/seo/{idea_id}/status": {
      "get": {
        "tags": [
          "SEO & Keywords"
        ],
        "summary": "Get Seo Status",
        "operationId": "get_seo_status_seo__idea_id__status_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/seo/{idea_id}": {
      "get": {
        "tags": [
          "SEO & Keywords"
        ],
        "summary": "Get Full Seo Metadata",
        "operationId": "get_full_seo_metadata_seo__idea_id__get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SEOMetadataSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/seo/{idea_id}/generate": {
      "post": {
        "tags": [
          "SEO & Keywords"
        ],
        "summary": "Generate Seo Metadata",
        "operationId": "generate_seo_metadata_seo__idea_id__generate_post",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          },
          {
            "name": "apply",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Apply"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          },
          {
            "name": "use_browser",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Use Browser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/pipeline/ideas/{idea_id}/progress": {
      "get": {
        "tags": [
          "Pipeline Orchestrator"
        ],
        "summary": "Get Idea Pipeline Progress",
        "operationId": "get_idea_pipeline_progress_pipeline_ideas__idea_id__progress_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pipeline/ideas/{idea_id}/run": {
      "post": {
        "tags": [
          "Pipeline Orchestrator"
        ],
        "summary": "Run Full Pipeline For Idea",
        "operationId": "run_full_pipeline_for_idea_pipeline_ideas__idea_id__run_post",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRunRequest",
                "default": {
                  "skip_browser": true,
                  "dry_run": true
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompetitorItemSchema": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "channel": {
            "type": "string",
            "title": "Channel",
            "default": ""
          },
          "video_id": {
            "type": "string",
            "title": "Video Id",
            "default": ""
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          },
          "view_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Count"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CompetitorItemSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineProgressResponse": {
        "properties": {
          "idea_id": {
            "type": "integer",
            "title": "Idea Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "stages": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stages"
          },
          "overall_status": {
            "type": "string",
            "title": "Overall Status"
          }
        },
        "type": "object",
        "required": [
          "idea_id",
          "title",
          "stages",
          "overall_status"
        ],
        "title": "PipelineProgressResponse"
      },
      "PipelineRunRequest": {
        "properties": {
          "skip_browser": {
            "type": "boolean",
            "title": "Skip Browser",
            "default": true
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          }
        },
        "type": "object",
        "title": "PipelineRunRequest"
      },
      "PipelineRunResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "idea_id": {
            "type": "integer",
            "title": "Idea Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "idea_id",
          "status",
          "stage",
          "message"
        ],
        "title": "PipelineRunResponse"
      },
...
      "SEOHarvestCompetitorsRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SEOHarvestCompetitorsRequest"
      },
      "SEOHarvestCompetitorsResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "ran": {
            "type": "boolean",
            "title": "Ran"
          },
          "competitors": {
            "items": {
              "$ref": "#/components/schemas/CompetitorItemSchema"
            },
            "type": "array",
            "title": "Competitors"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "query",
          "ran",
          "competitors",
          "count"
        ],
        "title": "SEOHarvestCompetitorsResponse"
      },
      "SEOHarvestSuggestRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "lang": {
            "type": "string",
            "title": "Lang",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SEOHarvestSuggestRequest"
      },
      "SEOHarvestSuggestResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "query",
          "suggestions",
          "count"
        ],
        "title": "SEOHarvestSuggestResponse"
      },
      "SEOMetadataSchema": {
        "properties": {
          "idea_id": {
            "type": "integer",
            "title": "Idea Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "seo_description": {
            "type": "string",
            "title": "Seo Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "pinned_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned Comment"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_real_seo": {
            "type": "boolean",
            "title": "Is Real Seo"
          }
        },
        "type": "object",
        "required": [
          "idea_id",
          "title",
          "seo_description",
          "tags",
          "status",
          "is_real_seo"
        ],
        "title": "SEOMetadataSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}